🐛 fix: 받은 편지의 에너지 적용 시 에너지 모션 추가
