+++ src/app/(route)/(monster)/component/cards/_ui/MonsterCounterBox.tsx
@@ -7,36 +7,10 @@ import CharacterCanvas, {
 } from '@components/CharacterCanvas';
 import CounterBox, { type CounterBoxProps } from '@components/CounterBox';
 
-import useCanvas from '@hooks/useCanvas';
-import useConfetti from '@hooks/useConfetti';
-
 type MonsterCounterBoxPropps = CharacterCanvasProps &
   CounterBoxProps & {
     clear?: boolean;
   };
-
-const ConfettiMap = {
-  mad: {
-    image: {
-      src: '/images/thunder0000.png',
-      width: 75,
-      height: 75,
-    },
-    start: 'center',
-    speed: 10,
-    particleNumber: 10,
-  },
-  sad: {
-    image: {
-      src: '/images/heart_active.webp',
-      width: 50,
-      height: 50,
-    },
-    start: 'center',
-    speed: 6,
-    particleNumber: 10,
-  },
-} as const;
 
 export default function MonsterCounterBox({
   width = 280,
@@ -50,15 +24,7 @@ export default function MonsterCounterBox({
   onCount,
   onComplete,
 }: MonsterCounterBoxPropps) {
-  const canvasRef = useCanvas(width, height);
-  const { addConfetti, destroyCanvas } = useConfetti(canvasRef);
-
   const [isMouseDown, setIsMouseDown] = useState(false);
-
-  const handleCount = (count: number) => {
-    onCount(count);
-    addConfetti(ConfettiMap[type]);
-  };
 
   const handleMouseDown = debounce(() => !clear && setIsMouseDown(true), 5000, {
     leading: true,
@@ -70,34 +36,24 @@ export default function MonsterCounterBox({
     trailing: true,
   });
 
-  if (clear) {
-    destroyCanvas();
-  }
-
   return (
-    <>
-      <CounterBox
-        startAt={startAt}
-        endAt={endAt}
-        onCount={handleCount}
-        onComplete={onComplete}
-        onMouseDown={handleMouseDown}
-        onMouseUp={handleMouseUp}
-      >
-        <CharacterCanvas
-          width={width}
-          height={height}
-          className="h-full w-full"
-          type={type}
-          status={clear || isMouseDown ? 'after' : 'before'}
-          items={items}
-          background={background}
-        />
-      </CounterBox>
-      <canvas
-        ref={canvasRef}
-        className="pointer-events-none absolute left-0 top-0 h-full w-full select-none"
+    <CounterBox
+      startAt={startAt}
+      endAt={endAt}
+      onCount={onCount}
+      onComplete={onComplete}
+      onMouseDown={handleMouseDown}
+      onMouseUp={handleMouseUp}
+    >
+      <CharacterCanvas
+        width={width}
+        height={height}
+        className="h-full w-full"
+        type={type}
+        status={clear || isMouseDown ? 'after' : 'before'}
+        items={items}
+        background={background}
       />
-    </>
+    </CounterBox>
   );
 }
