+++ src/app/(route)/(monster)/component/cards/constants.ts
@@ -1,0 +1,22 @@
+export const ConfettiMap = {
+  mad: {
+    image: {
+      src: '/images/thunder0000.png',
+      width: 50,
+      height: 50,
+    },
+    start: 'center',
+    speed: 7,
+    particleNumber: 6,
+  },
+  sad: {
+    image: {
+      src: '/images/heart_active.webp',
+      width: 50,
+      height: 50,
+    },
+    start: 'center',
+    speed: 6,
+    particleNumber: 10,
+  },
+} as const;
